scripts/requirements.txt
openai>=2.24.0
python-dotenv>=1.0.0
numpy>=1.21.0
Pillow>=10.0.0
pdfplumber>=0.11.0
beautifulsoup4>=4.12.0
requests>=2.31.0
